#250: Improved forgot email functionality
The forgot email functionality now also works for participants, and not only for admins. When a user requests an email, he/she will receive an overview of all the parties he/she participates in and the parties he/she created. Only the parties of the past week and future are given. As described in #250