Switched from LIKE to = to prevent suboptimal queries
The query will only match a single e-mailaddress. The use of LIKE in combination with underscores in e-mailadresses will be treated as a single wildcard character search, resulting in a range query rather than a ref query. Other than a performance hit, this could also lead to unwanted results if the wildcard happens to match a different address than intended.